Webhooks/requirements.txt
requests
python-dotenv
pyttsx3
threading
subprocess
platform